COMMERCE-1067 Inline BatchEngineTaskResourceRegistry
